Added logic to only remove files if the folder is there

+++ src/UNL/PersonInfo/Record.php
@@ -44,9 +44,11 @@ class UNL_PersonInfo_Record
 
     public function clear_images() {
         $path_to_save_location = dirname(dirname(dirname(__DIR__))) . '/data/person_images/' . $this->uid;
-        $tmp_files = array_diff(scandir($path_to_save_location), array('.','..'));
-        foreach ($tmp_files as $file) {
-            unlink($path_to_save_location . '/' . $file);
+        if (file_exists($path_to_save_location)) {
+            $tmp_files = array_diff(scandir($path_to_save_location), array('.','..'));
+            foreach ($tmp_files as $file) {
+                unlink($path_to_save_location . '/' . $file);
+            }
         }
     }
